changed VGG16 to work with APTOS dataset and use imagenet weights and also add testing code
The VGG16 classifier now uses the imagenet weights and train the last block in the CNN along with the fully connected layers.
The data generators are configured for the APTOS dataset.
Testing code has also been added which evaluates the model on all data generators and calculates accuracy, confusion matrix, and classification report.